Kernel: Added poll syscall support

+++ kernel/api/posix/mqueue.h
@@ -2,9 +2,9 @@
 
 struct mq_attr
 {
-    long long mq_flags;
-    long long mq_maxmsg;
-    long long mq_msgsize;
-    long long mq_curmsgs;
-    long long __reserved[4];
+    long mq_flags;
+    long mq_maxmsg;
+    long mq_msgsize;
+    long mq_curmsgs;
+    long __reserved[4];
 };
